requirements.txt
asgiref==3.7.2
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==4.2
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
filelock==3.13.1
Flask==3.0.0
Flask-Session==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
platformdirs==4.1.0
protobuf==4.21.12
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.25.0
Werkzeug==3.0.1
